Ability to add tags for selected image(s). Tags can be selected from the known list, or typed and and use the 'add' button. Potential impact on the tagger: the tag CSV file count vs tag count is now allowed to be different (from adding new tags). Tag editing is still limited to 'general' tags.

+++ src/web.py
@@ -178,9 +178,9 @@ def index():
 
 @bp.route('/api/selection', methods=["GET"])
 def current_selection():
-    print('current_selection')
+    #print('current_selection')
     selected_ids = request.args.getlist('selected_ids', type=int)
-    print(selected_ids)
+    #print(selected_ids)
     if len(selected_ids) == 0:
         return jsonify([])
     results = current_app.db.get_common_tags(selected_ids,0,0.0)
@@ -188,21 +188,35 @@ def current_selection():
 
 @bp.route('/api/applyTagChanges', methods=["GET"])
 def applyTagChanges():
-    print('applyTagChanges')
+    #print('applyTagChanges')
     image_ids = request.args.getlist('image_ids', type=int)
     tag_ids = request.args.getlist('tag_ids', type=int)
+    text_tags = request.args.getlist('text_tags')
 
     blah = current_app.db.get_common_tags(image_ids,0,0.0)
     old_tag_ids = [row["tag_id"] for row in blah]
     
-    print(f"old_tag_ids: {old_tag_ids}")
-    print(f"new_tag_ids: {tag_ids}")
+    #print(f"ATC old_tag_ids: {old_tag_ids}")
+    #print(f"ATC new_tag_ids: {tag_ids}")
     
     tags_to_delete = list(set(old_tag_ids) - set(tag_ids))
-    print(f'tags to delete: {tags_to_delete}')
+    #print(f'ATC tags to delete: {tags_to_delete}')
     
-    current_app.db.delete_tags(image_ids, tags_to_delete)
-    
+    if len(tags_to_delete) > 0:
+        current_app.db.delete_tags(image_ids, tags_to_delete)
+
+    tags_to_add = list(set(tag_ids) - set(old_tag_ids))
+    #print(f'ATC tags to add: {tags_to_add}')
+
+    if len(tags_to_add) > 0:
+        #newdb = FlaskImageDb(configs.db_path, sql_echo=configs.sql_echo)        
+        current_app.db.add_tags(image_ids, tags_to_add)
+        #newdb.close()
+
+    #print(f'ATC text tags: {text_tags}')
+    if len(text_tags) > 0:
+        current_app.db.add_possibly_new_tags(image_ids, text_tags, 0) # TODO last parameter is hardcoded as GENERAL
+        
 #    if len(selected_ids) == 0:
 #        return jsonify([])
 #    results = current_app.db.get_common_tags(selected_ids,0,0.0)
